requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
discord.py==2.3.2
distro==1.9.0
flake8==7.0.0
Flask==3.0.2
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
grpcio==1.62.0
grpcio-tools==1.62.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.23
langchain-openai==0.0.2
langchainhub==0.1.14
langsmith==0.0.87
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.0
marshmallow-sqlalchemy==1.0.0
mccabe==0.7.0
multidict==6.0.5
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.7.2
packaging==23.2
pluggy==1.4.0
portalocker==2.8.2
protobuf==4.25.3
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.2
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
qdrant-client==1.8.0
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.27
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.2
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
youtube-transcript-api==0.6.2
